Move GraphLocus::midpoint out of gen-models into views
Relocate the midpoint helper (and its tests) from gen-models/src/locus.rs
into src/views/gen_graph_widget.rs as a free function locus_midpoint, and
update the gen-python and gen-r callers to use it.
