+++ gen-models/src/locus.rs
@@ -52,81 +52,4 @@ impl GraphLocus {
 
         out
     }
-
-    /// The slice and local offset (0-based, within that slice's block) at the
-    /// midpoint of this locus's total sequence length. `None` for an empty locus.
-    pub fn midpoint(&self) -> Option<(GraphNodeSlice, usize)> {
-        let total: usize = self
-            .slices
-            .iter()
-            .map(|slice| slice.end - slice.start)
-            .sum();
-        if total == 0 {
-            return None;
-        }
-        let half = total / 2;
-        let mut consumed = 0;
-        for (index, slice) in self.slices.iter().enumerate() {
-            let len = slice.end - slice.start;
-            let is_last = index == self.slices.len() - 1;
-            if consumed + len > half || is_last {
-                return Some((*slice, slice.start + (half - consumed)));
-            }
-            consumed += len;
-        }
-        None
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use gen_core::{HashId, Strand};
-    use gen_graph::GraphNode;
-
-    use super::{GraphLocus, GraphNodeSlice};
-
-    fn node(id: u8, start: i64, end: i64) -> GraphNode {
-        GraphNode {
-            node_id: HashId([id; 32]),
-            sequence_start: start,
-            sequence_end: end,
-        }
-    }
-
-    fn slice(block: GraphNode, start: usize, end: usize) -> GraphNodeSlice {
-        GraphNodeSlice {
-            block,
-            start,
-            end,
-            strand: Strand::Forward,
-        }
-    }
-
-    #[test]
-    fn test_midpoint_single_slice_is_centered_within_it() {
-        let block = node(1, 0, 10);
-        let locus = GraphLocus {
-            slices: vec![slice(block, 2, 8)],
-        };
-        // length 6, half = 3, so local offset = start(2) + 3 = 5
-        assert_eq!(locus.midpoint(), Some((slice(block, 2, 8), 5)));
-    }
-
-    #[test]
-    fn test_midpoint_spans_into_second_slice() {
-        let first = node(1, 0, 10);
-        let second = node(2, 0, 10);
-        let locus = GraphLocus {
-            slices: vec![slice(first, 8, 10), slice(second, 0, 10)],
-        };
-        // total length 12, half = 6; first slice covers 2 bases (consumed=2),
-        // midpoint falls 4 bases into the second slice.
-        assert_eq!(locus.midpoint(), Some((slice(second, 0, 10), 4)));
-    }
-
-    #[test]
-    fn test_midpoint_of_empty_locus_is_none() {
-        let locus = GraphLocus { slices: vec![] };
-        assert_eq!(locus.midpoint(), None);
-    }
 }
